backend-requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-firestore==2.18.0
pydantic==2.9.2
markdown==3.7
python-dateutil==2.9.0.post0